2026-03-19-fastapi-basics/SSC/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
gunicorn==21.2.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-jose[cryptography]==3.3.0
email-validator==2.1.0
python-multipart==0.0.6
slowapi==0.1.9
pytz==2023.3
firebase-admin==6.8.0
pymongo[srv]==4.10.1
groq>=0.5.0
